:sparkles:1011

+++ 1011.c
@@ -4,16 +4,16 @@ int main(){
 	int n,i,j;
 	char m;
 	int a,b,c;
-	int arr[3];
+	long arr[3];
 	int k = 1;
 	scanf("%d",&n);
 	while(k<=n){
 		i=0;
 		j=0;
-		scanf("%d",&arr[i]);
+		scanf("%ld",&arr[i]);
 		while(m=getchar() == 32){
 			i++;
-			scanf("%d",&arr[i]);
+			scanf("%ld",&arr[i]);
 		}
 		if(arr[j++]+arr[j++]>arr[j]){
 			printf("Case #%d: true\n",k);
